refactor: Rename ProposerAdapter files for consistency
This commit refactors the naming of all `ProposerAdapter` contracts, interfaces, and test files to improve file organization and readability. The `ProposerAdapter` prefix is now used at the beginning of the filenames, which groups related files together when sorted alphabetically.

Additionally, the proposer adapter contracts and tests have been moved into a new `proposer/` subdirectory to better separate them from other types of strategy adapters.

All internal references, imports, and contract/interface declarations have been updated to reflect the new naming convention.

**Key Renames:**
-   `IProposerAdapterV1` -> `IProposerAdapterBaseV1`
-   `ERC20ProposerAdapterV1` -> `ProposerAdapterERC20V1`
-   `ERC721ProposerAdapterV1` -> `ProposerAdapterERC721V1`
-   `HatsProposerAdapterV1` -> `ProposerAdapterHatsV1`
-   The corresponding interfaces and test files have also been renamed and moved.
