+++ contracts/interfaces/decent/deployables/IProposerAdapterERC20V1.sol
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: AGPL-3.0
 pragma solidity ^0.8.30;
 
-import {IProposerAdapterV1} from "./IProposerAdapterV1.sol";
+import {IProposerAdapterBaseV1} from "./IProposerAdapterBaseV1.sol";
 
-interface IERC20ProposerAdapterV1 is IProposerAdapterV1 {
+interface IProposerAdapterERC20V1 is IProposerAdapterBaseV1 {
     function initialize(address token_, uint256 proposerThreshold_) external;
 
     function token() external view returns (address token);
