+++ contracts/mocks/MockProposerAdapter.sol
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.30;
 
-import {IProposerAdapterV1} from "../interfaces/decent/deployables/IProposerAdapterV1.sol";
+import {IProposerAdapterBaseV1} from "../interfaces/decent/deployables/IProposerAdapterBaseV1.sol";
 
-contract MockProposerAdapter is IProposerAdapterV1 {
+contract MockProposerAdapter is IProposerAdapterBaseV1 {
     mapping(address => bool) private _isProposer;
 
     function isProposer(
